+++ script_binding/python/ftk/ftk_icon_view.py
@@ -9,7 +9,6 @@ __version__ = '$Id: $'
 import ctypes
 
 import ftk_dll
-import ftk_constants
 import ftk_typedef
 import ftk_bitmap
 import ftk_widget
@@ -17,6 +16,7 @@ import ftk_widget
 # ftk_icon_view.h
 
 _FtkWidgetPtr = ctypes.POINTER(ftk_widget.FtkWidget)
+
 _FtkBitmapPtr = ctypes.POINTER(ftk_bitmap.FtkBitmap)
 
 _user_data_refs = [None]
@@ -69,7 +69,8 @@ _FtkIconViewItemPtr = ctypes.POINTER(FtkIconViewItem)
 ftk_icon_view_create = ftk_dll.function('ftk_icon_view_create',
         '',
         args=['parent', 'x', 'y', 'width', 'height'],
-        arg_types=[_FtkWidgetPtr, ctypes.c_int, ctypes.c_int, ctypes.c_int, ctypes.c_int],
+        arg_types=[_FtkWidgetPtr, ctypes.c_int, ctypes.c_int, ctypes.c_int,
+            ctypes.c_int],
         return_type=_FtkWidgetPtr,
         dereference_return=True,
         require_return=True)
@@ -77,7 +78,7 @@ ftk_icon_view_create = ftk_dll.function('ftk_icon_view_create',
 ftk_icon_view_set_item_size = ftk_dll.function('ftk_icon_view_set_item_size',
         '',
         args=['thiz', 'size'],
-        arg_types=[_FtkWidgetPtr, ctypes.c_uint],
+        arg_types=[_FtkWidgetPtr, ctypes.c_size_t],
         return_type=ctypes.c_int,
         check_return=True)
 
@@ -101,12 +102,12 @@ ftk_icon_view_get_count = ftk_dll.function('ftk_icon_view_get_count',
         '',
         args=['thiz'],
         arg_types=[_FtkWidgetPtr],
-        return_type=ctypes.c_uint)
+        return_type=ctypes.c_size_t)
 
 ftk_icon_view_remove = ftk_dll.function('ftk_icon_view_remove',
         '',
         args=['thiz', 'index'],
-        arg_types=[_FtkWidgetPtr, ctypes.c_uint],
+        arg_types=[_FtkWidgetPtr, ctypes.c_size_t],
         return_type=ctypes.c_int,
         check_return=True)
 
@@ -118,7 +119,8 @@ ftk_icon_view_add = ftk_dll.function('ftk_icon_view_add',
         check_return=True)
 
 _ftk_icon_view_get = ftk_dll.private_function('ftk_icon_view_get',
-        arg_types=[_FtkWidgetPtr, ctypes.c_uint, ctypes.POINTER(_FtkIconViewItemPtr)],
+        arg_types=[_FtkWidgetPtr, ctypes.c_uint,
+            ctypes.POINTER(_FtkIconViewItemPtr)],
         return_type=ctypes.c_int,
         check_return=True)
